requirements.txt
boto3==1.26.19
botocore==1.29.19
pandas==1.3.5
psycopg2==2.9.5
pyarrow==9.0.0
python-dotenv==0.21.0
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
SQLAlchemy==1.4.44
